+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 <!-- fix: A bug fix -->
 <!-- chore: Changes to the build process or auxiliary tools and libraries -->
 
+### 0.3.0
+- chore: Update iOS SDK to v4.26.0.
+- chore: Update Android SDK to v2.20.0.
+- add: Added ability to pause and unpause surveys.
+
 ### 0.2.1
 - fix: Aligned iOS version to 16.0 for the example app and .podspec.